fix key dupe and west/east portal block placement

+++ src/main/java/palamod/procedures/PortalblockbreakProcedure.java
@@ -41,18 +41,8 @@ public class PortalblockbreakProcedure {
 		x_core = x;
 		y_core = y;
 		z_core = z;
-		if (!CheckportalstructureProcedure.execute(world, z_core, y_core, x_core, angle_block, shiny_wood)) {
+		if (!CheckportalstructureProcedure.execute(world, z_core, y_core, x_core, angle_block, shiny_wood) && getBlockNBTLogic(world, BlockPos.containing(x_core, y_core, z_core), "portal_powered")) {
 			if ((getBlockNBTString(world, BlockPos.containing(x_core, y_core, z_core), "position")).equals("west")) {
-				world.destroyBlock(BlockPos.containing(x_core + 1, y_core, z_core + 1), false);
-				world.destroyBlock(BlockPos.containing(x_core + 1, y_core, z_core), false);
-				world.destroyBlock(BlockPos.containing(x_core + 1, y_core, z_core - 1), false);
-				world.destroyBlock(BlockPos.containing(x_core + 2, y_core, z_core + 1), false);
-				world.destroyBlock(BlockPos.containing(x_core + 2, y_core, z_core - 1), false);
-				world.destroyBlock(BlockPos.containing(x_core + 2, y_core, z_core), false);
-				world.destroyBlock(BlockPos.containing(x_core + 3, y_core, z_core + 1), false);
-				world.destroyBlock(BlockPos.containing(x_core + 3, y_core, z_core), false);
-				world.destroyBlock(BlockPos.containing(x_core + 3, y_core, z_core - 1), false);
-			} else if ((getBlockNBTString(world, BlockPos.containing(x_core, y_core, z_core), "position")).equals("east")) {
 				world.destroyBlock(BlockPos.containing(x_core - 1, y_core, z_core + 1), false);
 				world.destroyBlock(BlockPos.containing(x_core - 1, y_core, z_core), false);
 				world.destroyBlock(BlockPos.containing(x_core - 1, y_core, z_core - 1), false);
@@ -62,6 +52,16 @@ public class PortalblockbreakProcedure {
 				world.destroyBlock(BlockPos.containing(x_core - 3, y_core, z_core + 1), false);
 				world.destroyBlock(BlockPos.containing(x_core - 3, y_core, z_core), false);
 				world.destroyBlock(BlockPos.containing(x_core - 3, y_core, z_core - 1), false);
+			} else if ((getBlockNBTString(world, BlockPos.containing(x_core, y_core, z_core), "position")).equals("east")) {
+				world.destroyBlock(BlockPos.containing(x_core + 1, y_core, z_core + 1), false);
+				world.destroyBlock(BlockPos.containing(x_core + 1, y_core, z_core), false);
+				world.destroyBlock(BlockPos.containing(x_core + 1, y_core, z_core - 1), false);
+				world.destroyBlock(BlockPos.containing(x_core + 2, y_core, z_core + 1), false);
+				world.destroyBlock(BlockPos.containing(x_core + 2, y_core, z_core - 1), false);
+				world.destroyBlock(BlockPos.containing(x_core + 2, y_core, z_core), false);
+				world.destroyBlock(BlockPos.containing(x_core + 3, y_core, z_core + 1), false);
+				world.destroyBlock(BlockPos.containing(x_core + 3, y_core, z_core), false);
+				world.destroyBlock(BlockPos.containing(x_core + 3, y_core, z_core - 1), false);
 			} else if ((getBlockNBTString(world, BlockPos.containing(x_core, y_core, z_core), "position")).equals("north")) {
 				world.destroyBlock(BlockPos.containing(x_core + 1, y_core, z_core - 1), false);
 				world.destroyBlock(BlockPos.containing(x_core, y_core, z_core - 1), false);
@@ -121,6 +121,13 @@ public class PortalblockbreakProcedure {
 					_level.sendBlockUpdated(_bp, _bs, _bs, 3);
 			}
 		}
+	}
+
+	private static boolean getBlockNBTLogic(LevelAccessor world, BlockPos pos, String tag) {
+		BlockEntity blockEntity = world.getBlockEntity(pos);
+		if (blockEntity != null)
+			return blockEntity.getPersistentData().getBoolean(tag);
+		return false;
 	}
 
 	private static String getBlockNBTString(LevelAccessor world, BlockPos pos, String tag) {
